Updated set_acceptance_rate_restart

+++ npmc/move_class.py
@@ -39,7 +39,7 @@ class Move(object):
                 file.seek(-2, os.SEEK_CUR)
             last_line = file.readline().decode()
             self.num_moves += int(last_line[2*index+1])
-            self.num_accepted += int(last_line[2*index+1] * last_line[2*index+2])
+            self.num_accepted += int(int(last_line[2*index+1]) * int(last_line[2*index+2]))
 
 class CBMCRegrowth(Move):
     """A class that encapsulates a Configurationally Biased Regrowth move as outline by Siepmann et al. that inherits from the Move class.
